requirements.txt
albumentations==1.3.1
codecarbon==2.2.3
datasets[vision]~=2.14.0
evaluate==0.3.0
ipadic==1.0.0
jiwer==3.0.2
joblib==1.3.1
loguru==0.7.0
pandas>=1.4
nltk==3.8.1
optuna==3.3.0
Pillow==10.0.0
protobuf==4.23.4
pydantic==1.10.11
sacremoses==0.0.53


scikit-learn==1.3.0
sentencepiece==0.1.99
tqdm==4.65.0
werkzeug==2.3.6
xgboost==1.7.6
huggingface_hub>=0.16.4
requests==2.31.0
gradio==3.41.0
einops==0.6.1
invisible-watermark==0.2.0
packaging==23.1
# latest versions
tensorboard
peft==0.5.0
trl==0.7.2
tiktoken==0.5.1
transformers==4.34.1
accelerate==0.24.0
diffusers==0.21.4
bitsandbytes==0.41.0
# extras
rouge_score==0.1.2
py7zr==0.20.6